web-service/requirements.txt
requests
python-dotenv
flask
flask-sqlalchemy
flask-smorest
gunicorn
psycopg2-binary